+++ src/setBit.h
@@ -1,6 +1,8 @@
 #ifndef SET_BIT_H
 #define SET_BIT_H
 
+#include "__stdint.h"
+
 #ifdef TEST_SET_BIT
 #include "__stdlib.h"
 #include "__stdio.h"
@@ -8,6 +10,6 @@
 #include "returnSuccess.h"
 #endif
 
-void setBit(unsigned char* byte, int index);
+void setBit(unsigned char* byte, uintmax_t index);
 
 #endif
